+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
+## Version 0.0.5
+### update
+* unit test (MockMvc) analyzed a bit
+* GreetingControllerAnalyzedAgainTest, GreetingControllerTest 
+
+
 ## Version 0.0.4
 ### update 
 * integration test analyzed a bit
 * GreetingControllerITest, GreetingControllerAnalyzedAgainITest
+
 
 ## Version 0.0.3
 ### new